# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quiz API",
        "description": "API Documentation",
        "contact": {
            "name": "API Support",
            "email": "marko.mandic.engr@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get paginated list of products",
                "description": "Returns a paginated list of all products with their primary image",
                "operationId": "getProductsList",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 45
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "description": "Creates a new product and returns the created product",
                "operationId": "storeProduct",
                "requestBody": {
                    "description": "Product data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "title": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The title field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a specific product",
                "description": "Returns a specific product with its image and variants",
                "operationId": "getProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "description": "Updates a product and returns the updated product",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Product data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "title": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "description": "Deletes a product and returns a success message",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the product to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreProductRequest": {
                "required": [
                    "title",
                    "handle",
                    "price"
                ],
                "properties": {
                    "title": {
                        "description": "Product title",
                        "type": "string",
                        "example": "Ergonomic Office Chair"
                    },
                    "handle": {
                        "description": "URL-friendly slug for the product",
                        "type": "string",
                        "example": "ergonomic-office-chair"
                    },
                    "price": {
                        "description": "Product price",
                        "type": "number",
                        "format": "decimal",
                        "example": 199.99
                    },
                    "image_id": {
                        "description": "ID of the product's primary image",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdateProductRequest": {
                "properties": {
                    "title": {
                        "description": "Product title",
                        "type": "string",
                        "example": "Updated Office Chair"
                    },
                    "handle": {
                        "description": "URL-friendly slug for the product",
                        "type": "string",
                        "example": "updated-office-chair"
                    },
                    "price": {
                        "description": "Product price",
                        "type": "number",
                        "format": "decimal",
                        "example": 249.99
                    },
                    "image_id": {
                        "description": "ID of the product's primary image",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Image": {
                "required": [
                    "id",
                    "url"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the image",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    },
                    "url": {
                        "description": "URL of the image",
                        "type": "string",
                        "example": "https://example.com/images/chair.jpg"
                    },
                    "created_at": {
                        "description": "When the image was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-07T14:30:00Z"
                    },
                    "updated_at": {
                        "description": "When the image was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-07T14:30:00Z"
                    }
                },
                "type": "object"
            },
            "Product": {
                "required": [
                    "id",
                    "title",
                    "handle",
                    "price"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the product",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "title": {
                        "description": "Product title",
                        "type": "string",
                        "example": "Ergonomic Office Chair"
                    },
                    "handle": {
                        "description": "URL-friendly slug for the product",
                        "type": "string",
                        "example": "ergonomic-office-chair"
                    },
                    "price": {
                        "description": "Product price",
                        "type": "number",
                        "format": "decimal",
                        "example": 199.99
                    },
                    "image_id": {
                        "description": "ID of the product's primary image",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "When the product was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-07T14:30:00Z"
                    },
                    "updated_at": {
                        "description": "When the product was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-07T14:30:00Z"
                    },
                    "image": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Image"
                            }
                        ],
                        "nullable": true,
                        "description": "The product's primary image"
                    },
                    "variants": {
                        "description": "Product variants",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Variant"
                        }
                    }
                },
                "type": "object"
            },
            "Variant": {
                "required": [
                    "id",
                    "product_id",
                    "handle",
                    "price"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the variant",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440002"
                    },
                    "product_id": {
                        "description": "ID of the product this variant belongs to",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "handle": {
                        "description": "URL-friendly slug for the variant",
                        "type": "string",
                        "example": "ergonomic-office-chair-red"
                    },
                    "price": {
                        "description": "Variant price",
                        "type": "number",
                        "format": "decimal",
                        "example": 219.99
                    },
                    "created_at": {
                        "description": "When the variant was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-07T14:30:00Z"
                    },
                    "updated_at": {
                        "description": "When the variant was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-07T14:30:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}